use common fingerprinting function for openssl & ed25519 keys
