+++ lib/net/ssh/authentication/ed25519.rb
@@ -14,6 +14,7 @@ require 'rbnacl/hash'
 
 require 'base64'
 
+require 'net/ssh/pub_key'
 require 'net/ssh/transport/cipher_factory'
 require 'bcrypt_pbkdf' unless RUBY_PLATFORM == "java"
 
@@ -27,6 +28,7 @@ module ED25519
   end
 
   class PubKey
+    include Net::SSH::PubKey
     attr_reader :verify_key
 
     def initialize(data)
@@ -56,10 +58,6 @@ module ED25519
     def to_pem
       # TODO this is not pem
       ssh_type + Base64.encode64(@verify_key.to_bytes)
-    end
-
-    def fingerprint
-      @fingerprint ||= OpenSSL::Digest::MD5.hexdigest(to_blob).scan(/../).join(":")
     end
   end
